requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
distro==1.9.0
dotenv==0.9.9
filelock==3.19.1
firebase_admin==7.1.0
Flask==3.1.2
frozenlist==1.8.0
fsspec==2025.9.0
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
markovify==0.9.4
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
networkx==3.4.1
numpy==2.3.3
openai==2.1.0
packaging==25.0
pillow==11.3.0
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.33.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.21.0
